move add tests to library
